Create DataCollect.txt
Should be able to take in an object, with a String as it's key. 
Included functions:
insertDict(string, object); inserts objections into the Dictionary.
removeFirst(); removes the first object in the Dictionary.
removeKey(string); removes the object with the same string key.
removeAll(); clears the entire Dictionary.
findKey(string); finds if there is a matching key in the dictionary, if it doesn't exist, prints an error message, otherwise it will print the object.